rename pkg/resource
