+++ executor/server.go
@@ -29,9 +29,9 @@ import (
 	dcontext "github.com/hanfei1991/microcosm/pkg/context"
 	"github.com/hanfei1991/microcosm/pkg/deps"
 	"github.com/hanfei1991/microcosm/pkg/errors"
+	"github.com/hanfei1991/microcosm/pkg/externalresource"
 	"github.com/hanfei1991/microcosm/pkg/metadata"
 	"github.com/hanfei1991/microcosm/pkg/p2p"
-	"github.com/hanfei1991/microcosm/pkg/resource"
 	"github.com/hanfei1991/microcosm/pkg/serverutils"
 	"github.com/hanfei1991/microcosm/test"
 	"github.com/hanfei1991/microcosm/test/mock"
@@ -60,7 +60,7 @@ type Server struct {
 	metastore       metadata.MetaKV
 	p2pMsgRouter    p2pImpl.MessageRouter
 	discoveryKeeper *serverutils.DiscoveryKeepaliver
-	resourceBroker  *resource.Broker
+	resourceBroker  *externalresource.Broker
 }
 
 func NewServer(cfg *Config, ctx *test.Context) *Server {
@@ -175,7 +175,7 @@ func (s *Server) buildDeps(wid lib.WorkerID) (*deps.Deps, error) {
 	if err != nil {
 		return nil, err
 	}
-	err = deps.Provide(func() resource.Proxy {
+	err = deps.Provide(func() externalresource.Proxy {
 		return proxy
 	})
 	if err != nil {
@@ -344,7 +344,7 @@ func (s *Server) Run(ctx context.Context) error {
 	}
 
 	// TODO: make the prefix configurable later
-	s.resourceBroker = resource.NewBroker(s.cfg.Name, s.cfg.Name, s.cli)
+	s.resourceBroker = externalresource.NewBroker(s.cfg.Name, s.cfg.Name, s.cli)
 
 	s.p2pMsgRouter = p2p.NewMessageRouter(p2p.NodeID(s.info.ID), s.info.Addr)
