+++ lib/mock_master_impl.go
@@ -15,9 +15,9 @@ import (
 	"github.com/hanfei1991/microcosm/pb"
 	dcontext "github.com/hanfei1991/microcosm/pkg/context"
 	"github.com/hanfei1991/microcosm/pkg/deps"
+	"github.com/hanfei1991/microcosm/pkg/externalresource"
 	"github.com/hanfei1991/microcosm/pkg/metadata"
 	"github.com/hanfei1991/microcosm/pkg/p2p"
-	"github.com/hanfei1991/microcosm/pkg/resource"
 )
 
 type MockMasterImpl struct {
@@ -66,7 +66,7 @@ type masterParamListForTest struct {
 	MetaKVClient          metadata.MetaKV
 	ExecutorClientManager client.ClientsManager
 	ServerMasterClient    client.MasterClient
-	ResourceProxy         resource.Proxy
+	ResourceProxy         externalresource.Proxy
 }
 
 func (m *MockMasterImpl) Reset() {
@@ -85,7 +85,7 @@ func (m *MockMasterImpl) Reset() {
 			MetaKVClient:          m.metaKVClient,
 			ExecutorClientManager: m.executorClientManager,
 			ServerMasterClient:    m.serverMasterClient,
-			ResourceProxy:         resource.NewMockProxy(m.id),
+			ResourceProxy:         externalresource.NewMockProxy(m.id),
 		}
 	})
 	if err != nil {
