+++ lib/worker.go
@@ -16,9 +16,9 @@ import (
 	"github.com/hanfei1991/microcosm/pkg/clock"
 	dcontext "github.com/hanfei1991/microcosm/pkg/context"
 	derror "github.com/hanfei1991/microcosm/pkg/errors"
+	"github.com/hanfei1991/microcosm/pkg/externalresource"
 	"github.com/hanfei1991/microcosm/pkg/metadata"
 	"github.com/hanfei1991/microcosm/pkg/p2p"
-	"github.com/hanfei1991/microcosm/pkg/resource"
 )
 
 type Worker interface {
@@ -60,7 +60,7 @@ type BaseWorker interface {
 	MetaKVClient() metadata.MetaKV
 	UpdateStatus(ctx context.Context, status WorkerStatus) error
 	SendMessage(ctx context.Context, topic p2p.Topic, message interface{}) (bool, error)
-	Resource() resource.Proxy
+	Resource() externalresource.Proxy
 	// Exit should be called when worker (in user logic) wants to exit
 	// - If err is nil, it means worker exits normally
 	// - If err is not nil, it means worker meets error, and after worker exits
@@ -74,7 +74,7 @@ type DefaultBaseWorker struct {
 	messageHandlerManager p2p.MessageHandlerManager
 	messageSender         p2p.MessageSender
 	metaKVClient          metadata.MetaKV
-	resourceProxy         resource.Proxy
+	resourceProxy         externalresource.Proxy
 
 	masterClient *masterClient
 	masterID     MasterID
@@ -103,7 +103,7 @@ type workerParams struct {
 	MessageHandlerManager p2p.MessageHandlerManager
 	MessageSender         p2p.MessageSender
 	MetaKVClient          metadata.MetaKV
-	ResourceProxy         resource.Proxy
+	ResourceProxy         externalresource.Proxy
 }
 
 func NewBaseWorker(
@@ -308,7 +308,7 @@ func (w *DefaultBaseWorker) SendMessage(
 	return w.messageSender.SendToNode(ctx, w.masterClient.MasterNode(), topic, message)
 }
 
-func (w *DefaultBaseWorker) Resource() resource.Proxy {
+func (w *DefaultBaseWorker) Resource() externalresource.Proxy {
 	return w.resourceProxy
 }
