+++ pkg/externalresource/broker.go
@@ -1,4 +1,4 @@
-package resource
+package externalresource
 
 import (
 	"context"
