+++ pkg/externalresource/manager_test.go
@@ -1,4 +1,4 @@
-package resource
+package externalresource
 
 import "testing"
