+++ pkg/externalresource/utils.go
@@ -1,4 +1,4 @@
-package resource
+package externalresource
 
 import (
 	"context"
